+++ pkg/receive/handler.go
@@ -12,18 +12,14 @@ import (
 	stdlog "log"
 	"net"
 	"net/http"
-	"net/url"
 	"sort"
 	"strconv"
 	"sync"
 	"time"
 
-	extflag "github.com/efficientgo/tools/extkingpin"
 	"github.com/thanos-io/thanos/pkg/api"
 	statusapi "github.com/thanos-io/thanos/pkg/api/status"
-	"github.com/thanos-io/thanos/pkg/httpconfig"
 	"github.com/thanos-io/thanos/pkg/logging"
-	"github.com/thanos-io/thanos/pkg/promclient"
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
@@ -87,34 +83,23 @@ var (
 
 // Options for the web Handler.
 type Options struct {
-	Writer                   *Writer
-	ListenAddress            string
-	Registry                 *prometheus.Registry
-	TenantHeader             string
-	TenantField              string
-	DefaultTenantID          string
-	ReplicaHeader            string
-	Endpoint                 string
-	ReplicationFactor        uint64
-	ReceiverMode             ReceiverMode
-	Tracer                   opentracing.Tracer
-	TLSConfig                *tls.Config
-	DialOpts                 []grpc.DialOption
-	ForwardTimeout           time.Duration
-	RelabelConfigs           []*relabel.Config
-	TSDBStats                TSDBStats
-	LimitsConfig             *RootLimitsConfig
-	SeriesLimitSupported     bool
-	MaxPerTenantLimit        uint64
-	MetaMonitoringUrl        *url.URL
-	MetaMonitoringHttpClient *extflag.PathOrContent
-	MetaMonitoringLimitQuery string
-}
-
-// activeSeriesLimiter encompasses active series limiting logic.
-type activeSeriesLimiter interface {
-	QueryMetaMonitoring(context.Context, log.Logger) error
-	isUnderLimit(string, log.Logger) (bool, error)
+	Writer            *Writer
+	ListenAddress     string
+	Registry          *prometheus.Registry
+	TenantHeader      string
+	TenantField       string
+	DefaultTenantID   string
+	ReplicaHeader     string
+	Endpoint          string
+	ReplicationFactor uint64
+	ReceiverMode      ReceiverMode
+	Tracer            opentracing.Tracer
+	TLSConfig         *tls.Config
+	DialOpts          []grpc.DialOption
+	ForwardTimeout    time.Duration
+	RelabelConfigs    []*relabel.Config
+	TSDBStats         TSDBStats
+	Limiter           *limiter
 }
 
 // Handler serves a Prometheus remote write receiving HTTP endpoint.
@@ -125,13 +110,12 @@ type Handler struct {
 	options  *Options
 	listener net.Listener
 
-	mtx               sync.RWMutex
-	hashring          Hashring
-	peers             *peerGroup
-	expBackoff        backoff.Backoff
-	peerStates        map[string]*retryState
-	receiverMode      ReceiverMode
-	ActiveSeriesLimit activeSeriesLimiter
+	mtx          sync.RWMutex
+	hashring     Hashring
+	peers        *peerGroup
+	expBackoff   backoff.Backoff
+	peerStates   map[string]*retryState
+	receiverMode ReceiverMode
 
 	forwardRequests   *prometheus.CounterVec
 	replications      *prometheus.CounterVec
@@ -166,7 +150,7 @@ func NewHandler(logger log.Logger, o *Options) *Handler {
 			Max:    30 * time.Second,
 			Jitter: true,
 		},
-		limiter: newLimiter(o.LimitsConfig, registerer),
+		limiter: o.Limiter,
 		forwardRequests: promauto.With(registerer).NewCounterVec(
 			prometheus.CounterOpts{
 				Name: "thanos_receive_forward_requests_total",
@@ -214,11 +198,6 @@ func NewHandler(logger log.Logger, o *Options) *Handler {
 		h.replicationFactor.Set(float64(o.ReplicationFactor))
 	} else {
 		h.replicationFactor.Set(1)
-	}
-
-	h.ActiveSeriesLimit = NewNopSeriesLimit()
-	if h.options.SeriesLimitSupported {
-		h.ActiveSeriesLimit = NewActiveSeriesLimit(h.options, registerer, h.receiverMode, logger)
 	}
 
 	ins := extpromhttp.NewNopInstrumentationMiddleware()
@@ -438,7 +417,7 @@ func (h *Handler) receiveHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 	defer h.limiter.writeGate.Done()
 
-	under, err := h.ActiveSeriesLimit.isUnderLimit(tenant, tLogger)
+	under, err := h.limiter.HeadSeriesLimiter.isUnderLimit(tenant)
 	if err != nil {
 		level.Error(tLogger).Log("msg", "error while limiting", "err", err.Error())
 	}
@@ -551,141 +530,6 @@ func (h *Handler) receiveHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 	h.writeTimeseriesTotal.WithLabelValues(strconv.Itoa(responseStatusCode), tenant).Observe(float64(len(wreq.Timeseries)))
 	h.writeSamplesTotal.WithLabelValues(strconv.Itoa(responseStatusCode), tenant).Observe(float64(totalSamples))
-}
-
-// activeSeriesLimit implements activeSeriesLimiter interface.
-type activeSeriesLimit struct {
-	mtx                    sync.RWMutex
-	limit                  uint64
-	tenantCurrentSeriesMap map[string]float64
-
-	metaMonitoringURL    *url.URL
-	metaMonitoringClient *http.Client
-	metaMonitoringQuery  string
-
-	configuredTenantLimit prometheus.Gauge
-	limitedRequests       *prometheus.CounterVec
-	metaMonitoringErr     prometheus.Counter
-}
-
-func NewActiveSeriesLimit(o *Options, registerer prometheus.Registerer, r ReceiverMode, logger log.Logger) *activeSeriesLimit {
-	limit := &activeSeriesLimit{
-		limit:               o.MaxPerTenantLimit,
-		metaMonitoringURL:   o.MetaMonitoringUrl,
-		metaMonitoringQuery: o.MetaMonitoringLimitQuery,
-		configuredTenantLimit: promauto.With(registerer).NewGauge(
-			prometheus.GaugeOpts{
-				Name: "thanos_receive_tenant_head_series_limit",
-				Help: "The configured limit for active (head) series of tenants.",
-			},
-		),
-		limitedRequests: promauto.With(registerer).NewCounterVec(
-			prometheus.CounterOpts{
-				Name: "thanos_receive_head_series_limited_requests_total",
-				Help: "The total number of remote write requests that have been dropped due to active series limiting.",
-			}, []string{"tenant"},
-		),
-		metaMonitoringErr: promauto.With(registerer).NewCounter(
-			prometheus.CounterOpts{
-				Name: "thanos_receive_metamonitoring_failed_queries_total",
-				Help: "The total number of meta-monitoring queries that failed while limiting.",
-			},
-		),
-	}
-
-	limit.configuredTenantLimit.Set(float64(o.MaxPerTenantLimit))
-	limit.tenantCurrentSeriesMap = map[string]float64{}
-
-	// Use specified HTTPConfig to make requests to meta-monitoring.
-	httpConfContentYaml, err := o.MetaMonitoringHttpClient.Content()
-	if err != nil {
-		level.Error(logger).Log("msg", "getting http client config", "err", err.Error())
-	}
-
-	httpClientConfig, err := httpconfig.NewClientConfigFromYAML(httpConfContentYaml)
-	if err != nil {
-		level.Error(logger).Log("msg", "parsing http config YAML", "err", err.Error())
-	}
-
-	limit.metaMonitoringClient, err = httpconfig.NewHTTPClient(*httpClientConfig, "meta-mon-for-limit")
-	if err != nil {
-		level.Error(logger).Log("msg", "improper http client config", "err", err.Error())
-	}
-
-	return limit
-}
-
-// QueryMetaMonitoring queries any Prometheus Query API compatible meta-monitoring
-// solution with the configured query for getting current active (head) series of all tenants.
-// It then populates tenantCurrentSeries map with result.
-func (a *activeSeriesLimit) QueryMetaMonitoring(ctx context.Context, logger log.Logger) error {
-	c := promclient.NewWithTracingClient(logger, a.metaMonitoringClient, httpconfig.ThanosUserAgent)
-
-	vectorRes, _, err := c.QueryInstant(ctx, a.metaMonitoringURL, a.metaMonitoringQuery, time.Now(), promclient.QueryOptions{})
-	if err != nil {
-		a.metaMonitoringErr.Inc()
-		return err
-	}
-
-	level.Debug(logger).Log("msg", "successfully queried meta-monitoring", "vectors", len(vectorRes))
-
-	a.mtx.Lock()
-	defer a.mtx.Unlock()
-	// Construct map of tenant name and current HEAD series.
-	for _, e := range vectorRes {
-		for k, v := range e.Metric {
-			if k == "tenant" {
-				a.tenantCurrentSeriesMap[string(v)] = float64(e.Value)
-				level.Debug(logger).Log("msg", "tenant value queried", "tenant", string(v), "value", e.Value)
-			}
-		}
-	}
-
-	return nil
-}
-
-// isUnderLimit ensures that the current number of active series for a tenant does not exceed given limit.
-// It does so in a best-effort way, i.e, in case meta-monitoring is unreachable, it does not impose limits.
-// TODO(saswatamcode): Add capability to configure different limits for different tenants.
-func (a *activeSeriesLimit) isUnderLimit(tenant string, logger log.Logger) (bool, error) {
-	a.mtx.RLock()
-	defer a.mtx.RUnlock()
-	if a.limit == 0 || a.metaMonitoringURL.Host == "" {
-		return true, nil
-	}
-
-	// In such limiting flow, we ingest the first remote write request
-	// and then check meta-monitoring metric to ascertain current active
-	// series. As such metric is updated in intervals, it is possible
-	// that Receive ingests more series than the limit, before detecting that
-	// a tenant has exceeded the set limits.
-	v, ok := a.tenantCurrentSeriesMap[tenant]
-	if !ok {
-		return true, errors.New("tenant not in current series map")
-	}
-
-	if v >= float64(a.limit) {
-		level.Error(logger).Log("msg", "tenant above limit", "currentSeries", v, "limit", a.limit)
-		a.limitedRequests.WithLabelValues(tenant).Inc()
-		return false, nil
-	}
-
-	return true, nil
-}
-
-// nopSeriesLimit implements activeSeriesLimiter interface as no-op.
-type nopSeriesLimit struct{}
-
-func NewNopSeriesLimit() *nopSeriesLimit {
-	return &nopSeriesLimit{}
-}
-
-func (a *nopSeriesLimit) QueryMetaMonitoring(_ context.Context, _ log.Logger) error {
-	return nil
-}
-
-func (a *nopSeriesLimit) isUnderLimit(_ string, _ log.Logger) (bool, error) {
-	return true, nil
 }
 
 // forward accepts a write request, batches its time series by